fix(event-handler): remove polyfill and move it to index
